+++ data_provider/__init__.py
@@ -31,7 +31,7 @@
 
 from .base import BaseFetcher, DataFetcherManager
 from .efinance_fetcher import EfinanceFetcher
-from .akshare_fetcher import AkshareFetcher
+from .akshare_fetcher import AkshareFetcher, is_hk_stock_code
 from .tushare_fetcher import TushareFetcher
 from .pytdx_fetcher import PytdxFetcher
 from .baostock_fetcher import BaostockFetcher
@@ -49,6 +49,7 @@ __all__ = [
     'YfinanceFetcher',
     'is_us_index_code',
     'is_us_stock_code',
+    'is_hk_stock_code',
     'get_us_index_yf_symbol',
     'US_INDEX_MAPPING',
 ]